+++ DiscordBot/Commands/Slash/Twitch/Delete.cs
@@ -8,12 +8,12 @@ public class DeleteStreamerSlashCommand : ISlashCommands
     public string CommandName => "delete-streamer";
 
     private readonly DatabaseService _db;
-    private readonly IConfiguration _config;
+    private readonly IConfiguration _configuration;
 
     public DeleteStreamerSlashCommand(IConfiguration config, DatabaseService db)
     {
         _db = db;
-        _config = config;
+        _configuration = config;
     }
 
     public async Task RegisterCommandsAsync(DiscordSocketClient client)
@@ -23,11 +23,25 @@ public class DeleteStreamerSlashCommand : ISlashCommands
             .WithDescription("Remove a streamer from the list.")
             .AddOption("user", ApplicationCommandOptionType.User, "Discord user", true);
 
-        await client.Rest.CreateGuildCommand(cmd.Build(), ulong.Parse(_config["GuildID"]));
+        await client.Rest.CreateGuildCommand(cmd.Build(), ulong.Parse(_configuration["GuildID"]));
     }
 
     public async Task HandleCommand(SocketSlashCommand command, DiscordSocketClient client)
     {
+        var roleChecker = new RequiredRoles(_configuration);
+
+        if (!roleChecker.HasRequiredRole(executor))
+        {
+            var embedBuilder = new EmbedBuilder()
+                    .WithTitle("Permission Refusée")
+                    .WithDescription("Vous n'avez pas la permission d'utiliser cette commande.")
+                    .WithColor(Color.Red)
+                    .WithCurrentTimestamp();
+
+            await command.RespondAsync(embed: embedBuilder.Build(), ephemeral: true);
+            return;
+        }
+
         var user = (SocketGuildUser)command.Data.Options.First(o => o.Name == "user").Value;
 
         bool success = await _db.DeleteStreamerAsync(user.Id);
